@@ -15,63 +15,67 @@
   "author": "alexakis@imis.athena-innovation.gr",
   "license": "ISC",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-cli": "^6.6.5",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^12.0.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-2": "^6.5.0",
-    "babelify": "^7.2.0",
+    "babel": "^6.23.0",
+    "babel-cli": "^6.26.0",
+    "babel-eslint": "^8.0.0",
+    "babel-jest": "^21.0.2",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "babelify": "^7.3.0",
     "browserify-shim": "^3.8.14",
-    "envify": "^4.0.0",
+    "envify": "^4.1.0",
     "enzyme": "^2.9.1",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.3.0",
     "grunt": "^1.0.1",
-    "grunt-browserify": "^5.0.0",
+    "grunt-browserify": "^5.2.0",
     "grunt-cli": "^1.2.0",
     "grunt-contrib-clean": "^1.1.0",
     "grunt-contrib-concat": "^1.0.1",
     "grunt-contrib-copy": "^1.0.0",
     "grunt-contrib-sass": "^1.0.0",
-    "grunt-contrib-uglify": "^3.0.1",
+    "grunt-contrib-uglify": "^3.1.0",
     "grunt-contrib-watch": "^1.0.0",
-    "grunt-eslint": "^19.0.0",
-    "grunt-jsdoc": "^2.1.0",
-    "jest": "^20.0.4",
-    "jsdoc": "^3.4.3",
+    "grunt-eslint": "^20.1.0",
+    "grunt-jsdoc": "^2.1.1",
+    "jest": "^21.1.0",
+    "jsdoc": "^3.5.5",
     "redux-mock-store": "^1.3.0",
     "regenerator-runtime": "^0.11.0"
   },
   "dependencies": {
-    "flat": "^2.0.1",
-    "history": "^4.6.1",
-    "http-status-codes": "^1.1.6",
+    "flat": "^4.0.0",
+    "history": "^4.7.2",
+    "http-status-codes": "^1.3.0",
     "immutable": "^3.8.1",
-    "intl-messageformat": "^1.3.0",
+    "intl-messageformat": "^2.1.0",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
+    "ol": "^4.3.3",
     "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-dropzone": "^4.1.2",
-    "react-intl": "^2.3.0",
-    "react-redux": "^5.0.5",
-    "react-router-dom": "^4.1.2",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-dropzone": "^4.1.3",
+    "react-intl": "^2.4.0",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
     "react-router-redux": "^4.0.8",
     "react-select": "^1.0.0-rc.10",
-    "react-toastify": "^2.0.0-rc.3",
-    "react-transition-group": "^1.1.3",
+    "react-table": "^6.5.3",
+    "react-toastify": "^2.0.0",
+    "react-transition-group": "^2.2.0",
     "react-wiz": "git+https://github.com/DAIAD/react-wizard.git",
-    "reactstrap": "^4.6.2",
-    "redux": "^3.6.0",
+    "reactstrap": "^4.8.0",
+    "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
     "tether": "^1.4.0",
-    "url-search-params": "^0.9.0"
+    "url-search-params": "^0.10.0",
+    "victory": "^0.22.2"
   },
   "browserify": {
     "transform": [
