@@ -1,10 +1,11 @@
-const React = require('react');
-const PropTypes = require('prop-types');
-const ReactRedux = require('react-redux');
-const { NavLink } = require('react-router-dom');
-const Immutable = require('immutable');
+import * as React from 'react';
+import * as  PropTypes from 'prop-types';
+import { NavLink } from 'react-router-dom';
+import * as Immutable from 'immutable';
 
 import * as Roles from '../../model/role';
+import { StaticRoutes, DynamicRoutes, buildPath } from '../../model/routes';
+
 import SecureContent from '../helpers/secure-content';
 
 const Sections = {
@@ -16,20 +17,12 @@ const Sections = {
 };
 
 class Sidebar extends React.Component {
+
   constructor(props) {
     super(props);
     this.state = {
       expanded: new Immutable.Set(),
     };
-  }
-
-  hasRole(role) {
-    if (!role) {
-      return false;
-    }
-
-    let user = this.props.user;
-    return (user && user.profile.roles.indexOf(role) !== -1);
   }
 
   render() {
@@ -50,7 +43,7 @@ class Sidebar extends React.Component {
           <ul className="nav">
 
             <li className="nav-item">
-              <NavLink to={'/dashboard'} className="nav-link" activeClassName="active">
+              <NavLink to={StaticRoutes.Dashboard} className="nav-link" activeClassName="active">
                 {'Dashboard'}
               </NavLink>
             </li>
@@ -61,12 +54,12 @@ class Sidebar extends React.Component {
               </a>
               <ul className="nav-dropdown-items">
                 <li className="nav-item">
-                  <NavLink to={'/resource/explorer'} className="nav-link" activeClassName="active">
+                  <NavLink to={StaticRoutes.ResourceExplorer} className="nav-link" activeClassName="active">
                     <i className="fa fa-database"></i>{'Explorer'}
                   </NavLink>
                 </li>
                 <li className="nav-item">
-                  <NavLink to={'/resource/register'} className="nav-link" activeClassName="active">
+                  <NavLink to={StaticRoutes.ResourceRegistration} className="nav-link" activeClassName="active">
                     <i className="fa fa-pencil"></i>{'Register'}
                   </NavLink>
                 </li>
@@ -79,12 +72,12 @@ class Sidebar extends React.Component {
               </a>
               <ul className="nav-dropdown-items">
                 <li className="nav-item">
-                  <NavLink to={'/process/scheduler'} className="nav-link" activeClassName="active">
-                    <i className="fa fa-clock-o"></i>{'Scheduler'}
+                  <NavLink to={StaticRoutes.ProcessExplorer} className="nav-link" activeClassName="active">
+                    <i className="fa fa-clock-o"></i>{'ProcessExplorer'}
                   </NavLink>
                 </li>
                 <li className="nav-item">
-                  <NavLink to={'/process/design'} className="nav-link" activeClassName="active">
+                  <NavLink to={DynamicRoutes.ProcessDesignerCreate} className="nav-link" activeClassName="active">
                     <i className="fa fa-magic"></i>{'Design'}
                   </NavLink>
                 </li>
@@ -97,13 +90,8 @@ class Sidebar extends React.Component {
               </a>
               <ul className="nav-dropdown-items">
                 <li className="nav-item">
-                  <NavLink to={'/recipe/explorer'} className="nav-link" activeClassName="active">
+                  <NavLink to={StaticRoutes.RecipeExplorer} className="nav-link" activeClassName="active">
                     <i className="fa fa-book"></i>{'Overview'}
-                  </NavLink>
-                </li>
-                <li className="nav-item">
-                  <NavLink to={'/recipe/design'} className="nav-link" activeClassName="active">
-                    <i className="fa fa-magic"></i>{'Design'}
                   </NavLink>
                 </li>
               </ul>
@@ -115,35 +103,33 @@ class Sidebar extends React.Component {
               </a>
               <ul className="nav-dropdown-items">
                 <li className="nav-item">
-                  <NavLink to={'/tools/schema-mapper'} className="nav-link" activeClassName="active">
+                  <NavLink to={StaticRoutes.SchemaExplorer} className="nav-link" activeClassName="active">
                     <i className="fa fa-file-code-o"></i>{'Schema Mapper'}
                   </NavLink>
                 </li>
               </ul>
             </li>
 
-            {this.hasRole(Roles.ADMIN) &&
+            <SecureContent role={Roles.ADMIN}>
               <li className={'nav-item nav-dropdown ' + (expanded(Sections.Admin) ? 'open' : '')}>
                 <a className="nav-link nav-dropdown-toggle" onClick={() => (toggle(Sections.Admin), false)}>
                   {'Admin'}
                 </a>
                 <ul className="nav-dropdown-items">
                   <li className="nav-item">
-                    <NavLink to={'/admin/user-manager'} className="nav-link" activeClassName="active">
+                    <NavLink to={StaticRoutes.UserManager} className="nav-link" activeClassName="active">
                       <i className="fa fa-users"></i>{'Users'}
                     </NavLink>
                   </li>
-                  <SecureContent role={Roles.MAINTAINER}>
-                    <li className="nav-item">
-                      <NavLink to={'/admin/event-viewer'} className="nav-link" activeClassName="active">
-                        <i className="fa fa-heartbeat"></i>{'Event Log'}
-                      </NavLink>
-                    </li>
-                  </SecureContent>
+                  <li className="nav-item">
+                    <NavLink to={StaticRoutes.EventViewer} className="nav-link" activeClassName="active">
+                      <i className="fa fa-heartbeat"></i>{'Event Log'}
+                    </NavLink>
+                  </li>
                 </ul>
               </li>
-            }
-            
+            </SecureContent>
+
           </ul>
         </nav>
       </div>
@@ -157,18 +143,4 @@ Sidebar.propTypes = {
   }),
 };
 
-//
-// Wrap into a connected component
-//
-
-const mapStateToProps = (state) => {
-  return {
-    user: state.user
-  };
-};
-
-const mapDispatchToProps = null;
-
-Sidebar = ReactRedux.connect(mapStateToProps, mapDispatchToProps)(Sidebar);
-
-module.exports = Sidebar;
+export default Sidebar;
