@@ -1,13 +1,20 @@
-const React = require('react');
+import * as React from 'react';
 
+/**
+ * Application entry point
+ *
+ * @class Dashboard
+ * @extends {React.Component}
+ */
 class Dashboard extends React.Component {
+
   render() {
     return (
       <div className="animated fadeIn">
-        Hello Dashboard!
       </div>
     );
   }
+
 }
 
-module.exports = Dashboard;
+export default Dashboard;
