@@ -1,13 +1,14 @@
 const React = require('react');
 
 class Aside extends React.Component {
+
   render() {
     return (
       <aside className="aside-menu">
-        Aside Menu
       </aside>
     );
   }
+
 }
 
 module.exports = Aside;
