@@ -1,13 +1,20 @@
 import * as React from 'react';
 
+/**
+ * Browse application events
+ *
+ * @class EventViewer
+ * @extends {React.Component}
+ */
 class EventViewer extends React.Component {
+
   render() {
     return (
       <div className="animated fadeIn">
-        Event Viewer
       </div>
     );
   }
+
 }
 
 export default EventViewer;
