@@ -1,13 +1,20 @@
 import * as React from 'react';
 
+/**
+ * Browse and manage resources
+ *
+ * @class ResourceExplorer
+ * @extends {React.Component}
+ */
 class ResourceExplorer extends React.Component {
+
   render() {
     return (
       <div className="animated fadeIn">
-        Resource Explorer
       </div>
     );
   }
+
 }
 
 export default ResourceExplorer;
