@@ -1,19 +1,20 @@
 import * as React from 'react';
 
 /**
- * Component for managing job scheduling
- * 
- * @class Scheduler
+ * Browse and manage schema mapping definitions
+ *
+ * @class SchemaExplorer
  * @extends {React.Component}
  */
-class Scheduler extends React.Component {
+class SchemaExplorer extends React.Component {
+
   render() {
     return (
       <div className="animated fadeIn">
-        Scheduler
       </div>
     );
   }
+
 }
 
-export default Scheduler;
+export default SchemaExplorer;
