@@ -1,13 +1,20 @@
 import * as React from 'react';
 
+/**
+ * Browse and manage user accounts
+ *
+ * @class UserManager
+ * @extends {React.Component}
+ */
 class UserManager extends React.Component {
+
   render() {
     return (
       <div className="animated fadeIn">
-        User Manager
       </div>
     );
   }
+
 }
 
 export default UserManager;
