@@ -1,11 +1,11 @@
 const { createBrowserHistory } = require('history');
 
-const history = createBrowserHistory();
-
 // A properly formatted basename should have a leading slash, but no trailing slash
 const basename = '/workbench';
 
+const history = createBrowserHistory();
+
 module.exports = {
+  basename,
   history,
-  basename
 };
